+++ src/main/java/br/com/ia/problemadamochila/service/imp/MochilaServiceImp.java
@@ -27,7 +27,7 @@ public class MochilaServiceImp implements MochilaService {
     }
 
     @Override
-    public List<MochilaDTO> calculaFitnessDaPopulacao(List<MochilaDTO> populacao, Integer pesoMax) {
+    public List<MochilaDTO> calculaFitnessDaPopulacao(List<MochilaDTO> populacao, BigDecimal pesoMax) {
         populacao.stream().forEach((mochila) -> {
             mochila.setFitness(calculaFitnessMochila(mochila, pesoMax));
         });
@@ -40,11 +40,11 @@ public class MochilaServiceImp implements MochilaService {
      * fitness será 20% maior que o das mochilas no padrão O fitness será a
      * razão entre o valor e o peso
      */
-    private BigDecimal calculaFitnessMochila(MochilaDTO mochila, Integer pesoMax) {
+    private BigDecimal calculaFitnessMochila(MochilaDTO mochila, BigDecimal pesoMax) {
         BigDecimal fitness;
         BigDecimal peso = mochila.getPeso();
 
-        if (mochila.getPeso().compareTo(BigDecimal.valueOf(pesoMax)) == 1) {
+        if (mochila.getPeso().compareTo(pesoMax) == 1) {
             peso = mochila.getPeso().multiply(new BigDecimal("1.2"));
         }
